app/requirements.txt
Flask>=2.0
gunicorn>=20.0
google-cloud-pubsub>=2.5
google-cloud-secret-manager>=2.5
# Cloud SQL Connector and DB Driver
cloud-sql-python-connector>=1.0
pg8000
SQLAlchemy>=1.4